Add zones to the registry by default
When adding my (new!) Amplipi to home assistant, the sources worked but not the zones.  I found that changing this value before adding my amplipi entries fixed the problem for me.  Otherwise, the zones just stayed disabled forever.

It seems like the code is expecting the async_update to run and mark the zones as enabled, but it's never getting called when the zones are disabled (in my experience anyway)